app/pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]

[project]
name = "RESTful_API_Animals_DNK"
version = "0.0.1"
description = "RESTful API that allows storing animal genetic test data and calculating aggregated statistics. The API supports adding data, viewing records, and calculating statistics by animal species."
authors = [
    {name = "Timofey", email = "yakovishintimofey@gmail.com"}
]

dependencies = [
    "Django==5.1.4",
    "djangorestframework>=3.15.1",
    "drf-yasg==1.21.7",
    "django-filter==24.3"
]